backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Document Processing
PyMuPDF==1.23.8
python-docx==1.1.0

# AI/ML Libraries
numpy==1.26.2
scikit-learn==1.7.1
openai==1.3.7
# Using faiss-cpu with a more compatible version for Windows
faiss-cpu==1.7.2
sentence-transformers==2.2.2

# Database
sqlalchemy==2.0.23
alembic==1.13.0

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
requests==2.31.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Build dependencies
setuptools>=65.0.0
wheel>=0.38.0 